+++ Tests/LinuxMain.swift
@@ -11,6 +11,7 @@ import XCTest
 
 XCTMain([
     testCase(AttributesRuleTests.allTests),
+    testCase(ColonRuleTests.allTests),
     testCase(ConfigurationTests.allTests),
     testCase(CustomRulesTests.allTests),
     testCase(ExtendedNSStringTests.allTests),